changed folder structure, also added section on event driven execution, node has an internal registry and code is not always executed in the order it's written
