--- 03-Node-Basics/8-Event-Driven-Execution/app.js
@@ -1,46 +1,0 @@
-const http = require('http');
-const fs = require('fs');
-
-const server = http.createServer((req, res) => {
-    const url = req.url;
-    const method = req.method;
-    if(url === '/'){
-    res.write('<html>')
-    res.write('<head><title>Enter Message</title></head>')
-    res.write('<body><form action="/message" method="POST"><input type="text" name="message"><button type="submit">Send</button></form></body>');
-    res.write('</html>')
-    return res.end();
-
-    }
-    if(url === '/message' && method === 'POST'){
-
-        const body = [];
-        req.on('data', (chunk) =>{
-              console.log(chunk)
-           
-              body.push(chunk)
-        } )
-     
-         req.on('end', () => {
-       
-              const parsedBody = Buffer.concat(body).toString();
-               console.log(parsedBody)
-               const parseMessage = parsedBody.split('=')[1]
-
-                  fs.writeFileSync('message.txt', parseMessage );
-         });
-     
-        res.statusCode = 302;
-          res.setHeader('Location', '/')
-          return res.end()
-    }
-    res.setHeader('Content-Type', 'text/html')
-    res.write('<html>')
-    res.write('<head><title>My First Page With Node</title></head>')
-    res.write('<body><h1>Hello World From Node.js Server</h1></body>')
-    res.write('</html>')
-    res.end();
-})
-
-server.listen(3000)
-
